+++ packages/salesforcedx-apex-debugger/src/index.ts
@@ -18,8 +18,10 @@ export interface VscodeDebuggerMessage {
   message: string;
 }
 
-export interface ProxySettings {
-  url: string;
-  strictSSL: boolean;
-  auth: string;
+export interface WorkspaceSettings {
+  proxyUrl: string;
+  proxyStrictSSL: boolean;
+  proxyAuth: string;
+  connectionTimeoutMs: number;
+  setBreakpointTimeoutMs: number;
 }