+++ packages/salesforcedx-apex-debugger/test/unit/commands/stepCommands.test.ts
@@ -14,6 +14,7 @@ import {
   StepOutCommand,
   StepOverCommand
 } from '../../../src/commands';
+import { DEFAULT_CONNECTION_TIMEOUT_MS } from '../../../src/constants';
 
 describe('Step commands', () => {
   let sendRequestSpy: sinon.SinonStub;
@@ -39,11 +40,14 @@ describe('Step commands', () => {
       type: 'POST',
       url:
         'https://www.salesforce.com/services/debug/v41.0/step/07cFAKE?type=into',
+      timeout: DEFAULT_CONNECTION_TIMEOUT_MS,
       headers: {
-        'Content-Type': 'application/json',
+        'Content-Type': 'application/json;charset=utf-8',
         Accept: 'application/json',
-        Authorization: `OAuth 123`
-      }
+        Authorization: `OAuth 123`,
+        'Content-Length': 0
+      },
+      data: undefined
     };
 
     await requestService.execute(command);
@@ -63,11 +67,14 @@ describe('Step commands', () => {
       type: 'POST',
       url:
         'https://www.salesforce.com/services/debug/v41.0/step/07cFAKE?type=out',
+      timeout: DEFAULT_CONNECTION_TIMEOUT_MS,
       headers: {
-        'Content-Type': 'application/json',
+        'Content-Type': 'application/json;charset=utf-8',
         Accept: 'application/json',
-        Authorization: `OAuth 123`
-      }
+        Authorization: `OAuth 123`,
+        'Content-Length': 0
+      },
+      data: undefined
     };
 
     await requestService.execute(command);
@@ -87,11 +94,14 @@ describe('Step commands', () => {
       type: 'POST',
       url:
         'https://www.salesforce.com/services/debug/v41.0/step/07cFAKE?type=over',
+      timeout: DEFAULT_CONNECTION_TIMEOUT_MS,
       headers: {
-        'Content-Type': 'application/json',
+        'Content-Type': 'application/json;charset=utf-8',
         Accept: 'application/json',
-        Authorization: `OAuth 123`
-      }
+        Authorization: `OAuth 123`,
+        'Content-Length': 0
+      },
+      data: undefined
     };
 
     await requestService.execute(command);