+++ packages/salesforcedx-sobjects-faux-generator/test/integration/sObjectDescribe.test.ts
@@ -14,8 +14,12 @@ import {
 } from '../../src/describe/sObjectDescribe';
 import * as util from './integrationTestUtil';
 
+// The CustomObjects are all identical in terms of fields, just different ones to test batch
+// and multiple objects for testing describeGlobal
 const PROJECT_NAME = `project_${new Date().getTime()}`;
 const CUSTOM_OBJECT_NAME = 'MyCustomObject__c';
+const CUSTOM_OBJECT2 = 'MyCustomObject2__c';
+const CUSTOM_OBJECT3 = 'MyCustomObject3__c';
 const CUSTOM_FIELD_FULLNAME = CUSTOM_OBJECT_NAME + '.MyCustomField__c';
 const SIMPLE_OBJECT_DIR = path.join(
   'test',
@@ -26,6 +30,8 @@ const SIMPLE_OBJECT_DIR = path.join(
 );
 
 const sobjectdescribe = new SObjectDescribe();
+const MIN_CUSTOMOBJECT_NUM_FIELDS = 9;
+const CUSTOMOBJECT_NUMBERFIELD_PRECISION = 18;
 
 // tslint:disable:no-unused-expression
 describe('Fetch sObjects', function() {
@@ -68,13 +74,19 @@ describe('Fetch sObjects', function() {
   });
 
   it('Should be able to call describeGlobal', async function() {
+    const objs = [CUSTOM_OBJECT_NAME, CUSTOM_OBJECT2, CUSTOM_OBJECT3];
     const cmdOutput = await sobjectdescribe.describeGlobal(
       process.cwd(),
       SObjectCategory.CUSTOM,
       username
     );
-    expect(cmdOutput.length).to.be.equal(1);
-    expect(cmdOutput[0]).to.be.equal(CUSTOM_OBJECT_NAME);
+    expect(cmdOutput.length).to.be.equal(3);
+    expect(cmdOutput[0]).to.be.oneOf(objs);
+    expect(cmdOutput[1]).to.be.oneOf(objs);
+    expect(cmdOutput[2]).to.be.oneOf(objs);
+    expect(cmdOutput[0]).to.not.equal(cmdOutput[1]);
+    expect(cmdOutput[0]).to.not.equal(cmdOutput[2]);
+    expect(cmdOutput[1]).to.not.equal(cmdOutput[2]);
   });
 
   it('Should be able to call describeSObject on custom object', async function() {
@@ -85,12 +97,34 @@ describe('Fetch sObjects', function() {
     );
     expect(cmdOutput.name).to.be.equal(CUSTOM_OBJECT_NAME);
     expect(cmdOutput.custom).to.be.true;
-    expect(cmdOutput.fields.length).to.be.least(9);
+    expect(cmdOutput.fields.length).to.be.least(MIN_CUSTOMOBJECT_NUM_FIELDS);
     const customField = cmdOutput.fields[cmdOutput.fields.length - 1];
     expect(customField.custom).to.be.true;
-    expect(customField.precision).to.be.equal(18);
+    expect(customField.precision).to.be.equal(
+      CUSTOMOBJECT_NUMBERFIELD_PRECISION
+    );
     expect(customField.scale).to.be.equal(0);
     expect(customField.name).to.be.equal('MyCustomField__c');
+  });
+
+  it('Should be able to call describeSObjectBatch on custom objects', async function() {
+    const cmdOutput = await sobjectdescribe.describeSObjectBatch(
+      process.cwd(),
+      [CUSTOM_OBJECT_NAME, CUSTOM_OBJECT2, CUSTOM_OBJECT3],
+      0,
+      username
+    );
+    expect(cmdOutput[0].name).to.be.equal(CUSTOM_OBJECT_NAME);
+    expect(cmdOutput[0].custom).to.be.true;
+    expect(cmdOutput[0].fields.length).to.be.least(MIN_CUSTOMOBJECT_NUM_FIELDS);
+    const customField = cmdOutput[0].fields[cmdOutput[0].fields.length - 1];
+    expect(customField.name).to.be.equal('MyCustomField__c');
+
+    expect(cmdOutput[1].name).to.be.equal(CUSTOM_OBJECT2);
+    expect(cmdOutput[1].custom).to.be.true;
+    expect(cmdOutput[1].fields.length).to.be.least(MIN_CUSTOMOBJECT_NUM_FIELDS);
+
+    expect(cmdOutput[2].name).to.be.equal(CUSTOM_OBJECT3);
   });
 
   it('Should be able to call describeSObject on standard object', async function() {